requirements.txt
torch
pyyaml
opencv-python
lpips
pytorch-msssim
termcolor
einops
scikit-image
torchvision
torch-summary
tensorboardX
wandb